refactor: move CodeListWithMetadata to types folder
