+++ frontend/libs/studio-content-library/src/ContentLibrary/LibraryBody/pages/CodeListPage/index.ts
@@ -1,3 +1,4 @@
 export { CodeListPage } from './CodeListPage';
-export type { CodeListWithMetadata, CodeListData, CodeListPageProps } from './CodeListPage';
+export type { CodeListData, CodeListPageProps } from './CodeListPage';
 export type { CodeListIdSource, CodeListReference } from './types/CodeListReference';
+export type { CodeListWithMetadata } from './types/CodeListWithMetadata';
