+++ frontend/libs/studio-content-library/src/ContentLibrary/LibraryBody/pages/CodeListPage/types/CodeListWithMetadata.ts
@@ -1,0 +1,6 @@
+import type { CodeList } from '@studio/components-legacy';
+
+export type CodeListWithMetadata = {
+  codeList: CodeList;
+  title: string;
+};
